Rename project to neighbormodels   * Renamed files and directories   * Updated toolchain settings where project name was used   * Updated import statements in project modules   * Updated documentation
