--- README.md
@@ -1,12 +1,0 @@
-DataMaterials — Neighbors
-=========================
-
-> Counts the number of pairwise neighbors in a lattice within and between sub-lattices.
-
-License
--------
-
-[MIT][mit-license]
-
-[mit-license]: https://opensource.org/licenses/MIT
-
