+++ README.rst
@@ -1,0 +1,13 @@
+DataMaterials — neighbormodels
+==============================
+
+A tool for building crystal lattice models with distance-dependent neighbor interactions.
+
+
+License
+-------
+
+MIT_
+
+.. _MIT: https://opensource.org/licenses/MIT
+
