--- docs/datamaterials_neighbors.interactions.rst
@@ -1,7 +1,0 @@
-datamaterials\_neighbors.interactions module
-============================================
-
-.. automodule:: datamaterials_neighbors.interactions
-    :members:
-    :undoc-members:
-    :show-inheritance:
