--- docs/datamaterials_neighbors.neighbors.rst
@@ -1,7 +1,0 @@
-datamaterials\_neighbors.neighbors module
-=========================================
-
-.. automodule:: datamaterials_neighbors.neighbors
-    :members:
-    :undoc-members:
-    :show-inheritance:
