--- docs/datamaterials_neighbors.rst
@@ -1,17 +1,0 @@
-datamaterials\_neighbors package
-================================
-
-.. automodule:: datamaterials_neighbors
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-Submodules
-----------
-
-.. toctree::
-
-   datamaterials_neighbors.interactions
-   datamaterials_neighbors.neighbors
-   datamaterials_neighbors.structure
-
