--- docs/datamaterials_neighbors.structure.rst
@@ -1,7 +1,0 @@
-datamaterials\_neighbors.structure module
-=========================================
-
-.. automodule:: datamaterials_neighbors.structure
-    :members:
-    :undoc-members:
-    :show-inheritance:
