+++ docs/neighbormodels.interactions.rst
@@ -1,0 +1,7 @@
+neighbormodels.interactions module
+==================================
+
+.. automodule:: neighbormodels.interactions
+    :members:
+    :undoc-members:
+    :show-inheritance:
