+++ docs/neighbormodels.neighbors.rst
@@ -1,0 +1,7 @@
+neighbormodels.neighbors module
+===============================
+
+.. automodule:: neighbormodels.neighbors
+    :members:
+    :undoc-members:
+    :show-inheritance:
