+++ docs/neighbormodels.rst
@@ -1,0 +1,17 @@
+neighbormodels package
+======================
+
+.. automodule:: neighbormodels
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+
+   neighbormodels.interactions
+   neighbormodels.neighbors
+   neighbormodels.structure
+
