+++ docs/neighbormodels.structure.rst
@@ -1,0 +1,7 @@
+neighbormodels.structure module
+===============================
+
+.. automodule:: neighbormodels.structure
+    :members:
+    :undoc-members:
+    :show-inheritance:
