+++ neighbormodels/structure.py
